+++ domains/repositories/song_repository.go
@@ -3,6 +3,6 @@ package repositories
 import "github.com/Basabi-lab/lms/domains/models"
 
 type SongRepository interface {
-	GetByID(id uint) (*models.Song, error)
-	Create(song *models.Song) (uint, error)
+	GetByID(id uint64) (*models.Song, error)
+	Create(song *models.Song) (uint64, error)
 }
